+++ Documents/ReleaseNotes.md
@@ -6,6 +6,8 @@
     - `DDSPropertyAction`. New class to support property actions for data-driven styles.
 	- `DDSBuilder`.
 	  - `CoalesceGet`. Changed the method to accept a list of `DDSPropertyAction` for cases.
+	- `GeometryFilterStyleBuilder`.
+	  - Refactored building geometry filter style expressions.
 - `11.0.0-Preview-1.9`:
 - `Data-Driven Styles`. 
 	- `DDSBuilder`. Underlying internal expression builders have been made public.